Move tracklet refinement GUI out of main code
